requirements.txt
django-cms<3.5
djangocms-admin-style>=1.2,<1.3
django-treebeard>=4.0,<5.0

djangocms-text-ckeditor>=3.2.1
djangocms-link>=1.8
djangocms-style>=1.7
djangocms-googlemap>=0.5
djangocms-snippet>=1.9
djangocms-video>=2.0
djangocms-column>=1.6
easy_thumbnails
django-filer>=1.2
cmsplugin-filer>=1.1
Django<1.9
pytz
django-classy-tags>=0.7
html5lib>=0.999999,<0.99999999
Pillow>=3.0
django-sekizai>=0.9
django-select2<5.0six
djangorestframework==3.4.6
mysqlclient==1.3.7
django-page-cms==1.9.6
django-taggit==0.21.2
django-taggit-autosuggest==0.3.0
django-allauth==0.27.0
django-templatetag-handlebars==1.3.1
decorator==3.4.0
django-appconf==1.0.2
django-formtools==1.0
django-mptt==0.8.6
django-polymorphic==0.8.1
djangocms-attributes-field==0.1.1
django-crum==0.7.1
django_modern_rpc==0.4.2
moment==0.6
requests==2.12.5
sqlalchemy>=1.0.14
celery==4.0
django_celery_beat==1.0.1
django_celery_results==1.0.1
django-invitations==1.8